Support masking layers from Figma
On the Rust side, parse the "is_mask" parameter from Figma from frames and vectors that have this property, and export this into the corresponding ViewShape that we write into a serialized Figma file.

In Kotlin, create a few data structures to help manager mask data and the nodes that they affect:

MaskManager keeps tracks of masks and nodes that use the mask. This includes all sibling nodes rendered after the mask until we hit another mask or a node with clip contents on. The mask also applies to children of siblings affected by the mask. The mask manager is a state object, so when something in the mask hierarchy changes, such as nodes affected by a mask being added or removed or the rendering order changes so that masks affect different nodes, a recompose is triggered. If the internals of a mask node changes, the mask manager also causes a recompose.

The MaskManager is created in DesignView() and passed to child DesignViews, DesignFrame, and the DesignFrame render function. During render, if we are rendering a mask, instead of rendering anything to the canvas, we add the mask data to the mask manager. If we are not rendering a mask, we ask our mask manager for a list of masks that affect the node. If there are any masks, we create a separate canvas for rendering so that we can render the node with its contents masked into a bitmap with alpha, then render the bitmap onto the main canvas. To render the node with masking, we do the following:

1. Combine the mask fills and strokes into a single path
2. Call clipPath() on the combined mask path to clip the node rendering based on the mask path
3. Render the node's path
4. Render the mask fills with blendmode DstIn to only render alpha where it overlaps with the node's path
5. Render the mask strokes with blendmode DstIn
6. Render the bitmap to the main canvas

Fixes #63